ml/requirements.txt
# ml/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
inference-sdk
python-multipart